mock generateApplicationId  function
